LevelFactory added doesn't create discs yet.

+++ QBert/LevelFactory.h
@@ -14,7 +14,7 @@ namespace qbert
 	class LevelFactory
 	{
 	public:
-		static const std::weak_ptr<Level> CreateLevel(dae::Scene& scene, LevelType type, int discAmount = 2, float scale = 1.0f, int levelSize = 7);
+		static const std::weak_ptr<Level> CreateLevel(dae::Scene& scene, LevelType type, int discAmount = 2, int scale = 1, int levelSize = 7);
 
 	private:
 		static int m_CubeHeightOffset;
